EX1+Pictures-DRF_and_CRUD_for_Smart_Home/requirements.txt
django
djangorestframework
psycopg2-binary

pillow